Fix a few issues in the text buffer. Seems to work fine now!
except that it crashes whenever you do anything except typing text.

This changes need to be talked about with the infinity maintainer,
I'm not sure you can use the inf buffer to calculate the offsets:
The problem which needed to be solved was knowing how many (and which)
characters have been removed from the document; KTE just emits
textremoved(range) and deletes the characters from the document
*before* emitting that signal, so there's no way to retrieve it.
This commit takes them from the infinity textbuffer instead. I'm not
sure if that's safe.