unit tests: add a designated wait function
This is not good -- but how else to wait for synchronization?
As far as I know infinity doesn't give information about whether a peer
has received an edit command (which makes sense but is bad in this case),
and just waiting for the two documents to have the same text doesn't
work either, since it might not catch erroneous additional operations.
So for now, just wait a fixed time (50ms), then compare.