Notify KDirNotify of changes if nodes are added or removed.
TODO: what URL do we use for the notification?
If e.g. the URL contains the port, then views which are for the URL
without the port (using the standard port) won't be updated.
i.e. for each change, there's a whole class of URLs which need to be
reloaded; if a file is added in the root directory, those will all
need to be reloaded:
 inf://localhost/
 inf://foo@localhost/
 inf://localhost:6523/
 inf://foo:bar@localhost/
etc. To me, it is unclear how this can be solved properly at the moment.