Framework to add actions to the GUI (such as "Disconnect", etc)
There is a slight complication, since the KXMLGuiClient instance which
creates our actions must be present the first time kate displays the view.
To achieve this, we're forced to create it immediately when addView()
is called. However, at that point we do not usually know if a document
is a collaborative document or not.
Thus, we have to create a view in either way, and construct the XMLGUI
client immediately. In the future, it will be necessary to use
the "action lists" feature of KXMLGui to only display the GUI if
the document actually is a collaborative document.