remove localTextChanged()
There's no point in this, it's just a signal which is emitted when either
textInserted() or textRemoved() happens, and thus it does not need
to be handled seperately.
For example, if you select text in kate and paste something "over"
that text, it will still first emit textRemoved(), then textInserted()
(not only textChanged()).