Implement "open collaborative document"
This is not great, since it will replace the currently open document
instead of creating a new one. The latter is not possible from a KTE
plugin unfortunately. However the use case "launch editor -> connect
to some collab server" is still well covered by this, and in the end
we want to go for telepathy-based document hosting anyways.