Use the (not yet public) KTE::BufferInterface for buffer manipulation
This fixes bugs such as the space/tab replacement bug by using functions
in KTE which insert text into the buffer without any post-processing.
The interface required for this, KTextEditor::BufferInterface, will
be proposed for review later.
Since the KTE interface is not currently public, the old method for
inserting text will be used if it is not available. The same holds
in case the used KTE implementation does not implement the BufferInterface.