Severely reduce the amount of MovingRange instances used for highlighting.
Until now, the background highlighting just created a new MovingRange for
each chunk (usually character) inserted by a user. While convenient
code-wise, this is bad since it will a) use lots of memory (a few hundred
bytes *per character*) and b) require KTextEditor to track thousands
of ranges (which made e.g. kate visibly sluggish already when holding
down a key).
This commit introduces a basic algorithm for re-using ranges; it will
expand a range at the beginning or at the end if the color matches
the color of the newly inserted text, and just do nothing if the
text inserted is already contained within a range of the same color. For
my simple test documents, this reduces the amount of ranges from "a couple
of thousands" to "about 10".

A future optimization would be to merge ranges with the same color again
if the text between them is removed. It's questionable whether that is
worth it, though.