Re-arrange text buffer synchronization
There was bugs in the text buffer synchronization (synchronization
between the infinity textbuffer and kate's textbuffer) related to
newlines. Previous commits added tests which exposed those bugs, and this
one fixes the reason(s) for the issue -- which was actually a variety of
reasons, including incorrect (line, column) -> offset translation
functions and using the wrong translation function.

It also uses the signature of the textRemoved signal which tells about
the text which was removed (not only the range), which makes things simpler
there.

One thing is broken in kate, see the test marked as XFAIL for more
information.