Remove buffer interface pointer from the textbuffer header file
This caused a funny memory corruption, since if you #included the header
file without the #define being set, you would end up with a class size
which was too small. Thus, when using new, not enough memory was allocated
to hold all the members of the class, which ultimately led to crashes.