Export a fake TCP socket for StreamTubeServer when creating it.
The server will only register itself on dbus as soon as this happens.
What we can do is exporting a fake socket, and then waiting for the
tubeRequested signal. That signal is emitted shortly before the tube
is bound to the selected port, so in that slot we can set the real
one.